Remove some casts that are now unnecessary

+++ zlib-rs/src/deflate/algorithm/medium.rs
@@ -80,7 +80,7 @@ pub fn deflate_medium(stream: &mut DeflateStream, flush: DeflateFlush) -> BlockS
                     crate::deflate::longest_match::longest_match(state, hash_head);
                 state.match_start = match_start;
                 current_match.match_length = match_length as u16;
-                current_match.match_start = match_start as u16;
+                current_match.match_start = match_start;
                 if (current_match.match_length as usize) < WANT_MIN_MATCH {
                     current_match.match_length = 1;
                 }
@@ -123,7 +123,7 @@ pub fn deflate_medium(stream: &mut DeflateStream, flush: DeflateFlush) -> BlockS
                     crate::deflate::longest_match::longest_match(state, hash_head);
                 state.match_start = match_start;
                 next_match.match_length = match_length as u16;
-                next_match.match_start = match_start as u16;
+                next_match.match_start = match_start;
 
                 if next_match.match_start >= next_match.strstart {
                     /* this can happen due to some restarts */
